requirements.txt
aiolimiter==1.1.0
anyio==4.4.0
APScheduler==3.10.4
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
cryptography==42.0.8
exceptiongroup==1.2.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
pycparser==2.22
python-dotenv==1.0.1
python-telegram-bot==21.3
pytz==2024.1
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.31
tomli==2.0.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
